+++ packages/bootstrap/extra/dcos_internal_utils/cli.py
@@ -241,6 +241,11 @@ def dcos_cockroach_config_change(b, opts):
     _create_private_directory(path=cockroach_tmpdir, owner=user)
 
 
+@check_root
+def dcos_cluster_id(b, opts):
+    b.cluster_id()
+
+
 def noop(b, opts):
     return
 
@@ -269,6 +274,7 @@ bootstrappers = {
     'dcos-telegraf-master': dcos_telegraf_master,
     'dcos-telegraf-agent': dcos_telegraf_agent,
     'dcos-ui-update-service': noop,
+    'dcos-cluster-id': dcos_cluster_id,  # used for testing
 }